// generated/src/app/api-models/prov-specialty.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProvSpecialty {
  seqProvId: number;
  specialtyType: string;
  primarySpecialty: string;
  boardStatus: string;
  directoryInclude: string;
  acceptsPatients: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}